make form WD daemon

+++ dev/wd/subprocess_handler.py
@@ -1,47 +1,58 @@
 import os
 import subprocess
-import sys
 import signal
 import threading
 import time
+import sys
 
 
 class Subprocess(threading.Thread):
     def __init__(self):
         threading.Thread.__init__(self)
-        self.daemon = True
+        self.pipe_id = 0;
+        if os.path.exists('.lock'):
+            os.remove('.lock')
 
     def run(self):
         while True:
-            time.sleep(1)
             if os.path.exists('.lock'):
-                self.pipe.wait()
-                os.remove('.lock')
+                if not str(self.pipe.poll()) == "None":
+                    os.remove('.lock')
+                    self.start_subprocess()
+            time.sleep(1)
 
     def start_subprocess(self):
         if os.path.exists('.lock'):
-            return "Subprocess is already started. PID: {0}".format(self.pipe.pid)
+            return "Subprocess is already started. PID: {0}".format(self.pipe_id)
         else:
             self.file = "./../subprocess/subprocess.py"
             self.pipe = subprocess.Popen([sys.executable, self.file])
+            self.pipe_id = self.pipe.pid
             self.lock_file = open('.lock', 'w')
-            return "Subprocess successfully started. PID: {0}".format(self.pipe.pid)
+            return "Subprocess successfully started. PID: {0}".format(self.pipe_id)
 
     def stop_subprocess(self):
         if os.path.exists('.lock'):
             try:
-                os.kill(self.pipe.pid, signal.SIGTERM)
+                os.kill(self.pipe_id, signal.SIGTERM)
                 self.pipe.wait(3)
             except subprocess.TimeoutExpired:
-                os.kill(self.pipe.pid, signal.SIGKILL)
+                os.kill(self.pipe_id, signal.SIGKILL)
+            os.remove('.lock')
             return "Subprocess stopped"
         else:
             return "Subprocess not started"
+
+    def about_subprocess(self):
+        if os.path.exists('.lock'):
+            return "Subprocess is started. PID: {0}".format(self.pipe_id)
+        else:
+            return "Subprocess does not work"
 
     def execute_command(self, command):
         if command == "start":
             return self.start_subprocess()
         elif command == "stop":
             return self.stop_subprocess()
-        elif command == "status":
-            return "Subprocess is started" if os.path.exists('.lock') else "Subprocess does not work"
+        elif command == "about":
+            return self.about_subprocess()
